+++ packages/core/src/Core/Model/Company/Company.php
@@ -33,7 +33,6 @@ class Company
      */
     private $address;
 
-
     /**
      * @var string
      */